feat(back): create network to docker services, add user service and more
