+++ backend/.env.example
@@ -4,13 +4,8 @@ NODE_ENV=development
 APP_KEY=TgEmJYxy0KRBtW0FAnDqjJFY7cZq5hZT
 
 DB_CONNECTION=mysql
-DB_HOST=warren-mysql
-DB_PORT=3306
-DB_NAME=warren
-DB_USER=warren
-DB_PASSWORD=warren
-MYSQL_HOST=
-MYSQL_PORT=
-MYSQL_USER=
-MYSQL_PASSWORD=
-MYSQL_DB_NAME=
+MYSQL_HOST=warren-mysql
+MYSQL_PORT=3306
+MYSQL_USER=warren
+MYSQL_PASSWORD=warren
+MYSQL_DB_NAME=warren
