+++ backend/app/Schemas/UserSchema.ts
@@ -1,0 +1,6 @@
+export interface UserSchema {
+  id?: string | null
+  name: string
+  email: string
+  password?: string | null
+}
